# src/index.ts
import { serve } from "@hono/node-server";
import { Hono } from "hono";
import { cors } from "hono/cors";
import { logger } from "hono/logger";
import { serveStatic } from "@hono/node-server/serve-static";
import { randomBytes } from "crypto";
import { v4 as uuidv4 } from "uuid";
import {
  runMigrations,
  sqlite,
  getAgentByKey,
  getAgentById,
  createAgent,
  creditAgent,
  debitAgent,
  getDomainByName,
  getDomainsByAgent,
  getDomainForAgent,
  insertDomain,
  insertReferralEarning,
  getReferralEarnings,
  getReferralCount,
  getReferralDomainCount,
  type Agent,
} from "./db.js";
import {
  TLD_PRICES,
  getTldPrice,
  getTldFromDomain,
  getDomain as njallaGetDomain,
  listDomains as njallaListDomains,
  registerDomain as njallaRegisterDomain,
  listRecords as njallaListRecords,
  addRecord as njallaAddRecord,
  removeRecord as njallaRemoveRecord,
} from "./njalla.js";

// ─── Startup ───

runMigrations();

// ─── Rate limiter ───

const rateLimitBuckets = new Map<string, { count: number; windowStart: number }>();
setInterval(() => {
  const cutoff = Date.now() - 120_000;
  for (const [key, bucket] of rateLimitBuckets) {
    if (bucket.windowStart < cutoff) rateLimitBuckets.delete(key);
  }
}, 300_000);

function rateLimit(maxRequests: number, windowMs: number) {
  return async (c: any, next: () => Promise<void>) => {
    const ip =
      c.req.header("x-forwarded-for")?.split(",")[0]?.trim() ||
      c.req.header("x-real-ip") ||
      "unknown";
    const key = `${c.req.path}:${ip}`;
    const now = Date.now();
    const bucket = rateLimitBuckets.get(key);
    if (!bucket || now - bucket.windowStart > windowMs) {
      rateLimitBuckets.set(key, { count: 1, windowStart: now });
    } else {
      bucket.count++;
      if (bucket.count > maxRequests) {
        return c.json(
          { error: "rate_limited", message: `Too many requests. Limit: ${maxRequests} per ${windowMs / 1000}s` },
          429
        );
      }
    }
    await next();
  };
}

// ─── Auth middleware ───

async function requireAuth(c: any, next: () => Promise<void>) {
  const authHeader = c.req.header("Authorization");
  if (!authHeader?.startsWith("Bearer ")) {
    return c.json(
      { error: "unauthorized", message: "Missing Authorization header. Use: Bearer <api_key>" },
      401
    );
  }
  const apiKey = authHeader.slice(7);
  const agent = getAgentByKey(apiKey);
  if (!agent) {
    return c.json({ error: "unauthorized", message: "Invalid API key" }, 401);
  }
  c.set("agent", agent);
  c.set("agentId", agent.id);
  await next();
}

// ─── Domain validation ───

function isValidDomain(domain: string): boolean {
  const parts = domain.split(".");
  if (parts.length < 2) return false;
  const domainRegex = /^[a-zA-Z0-9]([a-zA-Z0-9-]*[a-zA-Z0-9])?$/;
  return parts.every((p) => p.length > 0 && domainRegex.test(p));
}

// ─── App setup ───

const app = new Hono<{ Variables: { agent: Agent; agentId: string } }>();

app.use("*", cors());
app.use("*", logger());

// ─── Static files ───

app.use("/llms.txt", serveStatic({ path: "public/llms.txt" }));
app.use("/openapi.json", serveStatic({ path: "public/openapi.json" }));

// ─── Health ───

app.get("/health", (c) =>
  c.json({ status: "ok", service: "agent-domains", version: "1.0.0" })
);

// ─── Root ───

app.get("/", (c) =>
  c.json({
    service: "Purple Flea Agent Domains",
    version: "1.0.0",
    description: "AI agents can register, manage, and configure domains with crypto payments.",
    docs: "/openapi.json",
    llms: "/llms.txt",
    endpoints: {
      public: ["GET /health", "GET /tlds", "GET /search", "POST /register"],
      wallet: ["POST /wallet/deposit", "GET /wallet/balance"],
      domains: [
        "POST /domains/purchase",
        "GET /domains",
        "GET /domains/:domain",
        "GET /domains/:domain/records",
        "POST /domains/:domain/records",
        "PUT /domains/:domain/records/:id",
        "DELETE /domains/:domain/records/:id",
      ],
      referrals: ["GET /referrals"],
    },
  })
);

// ─── POST /register ───

app.use("/register", rateLimit(10, 60_000));
app.post("/register", async (c) => {
  try {
    const body = await c.req.json().catch(() => ({}));
    const referralCode = body.referral_code as string | undefined;

    const agentId = uuidv4();
    const apiKey = `sk_domains_${randomBytes(24).toString("hex")}`;

    let referrerId: string | undefined;
    if (referralCode) {
      const referrer = getAgentById(referralCode);
      if (referrer) referrerId = referrer.id;
    }

    createAgent(agentId, apiKey, referrerId);

    return c.json(
      {
        api_key: apiKey,
        agent_id: agentId,
        message:
          "Store your API key securely — it cannot be recovered. Use Authorization: Bearer <api_key> on all requests.",
      },
      201
    );
  } catch (err: any) {
    return c.json({ error: "registration_failed", message: err.message }, 500);
  }
});

// ─── GET /tlds ───

app.get("/tlds", (c) => {
  const tlds = Object.entries(TLD_PRICES).map(([tld, price]) => ({
    tld: `.${tld}`,
    price_usdc: price,
    price_per_year: true,
  }));
  return c.json({ tlds });
});

// ─── GET /search ───

app.get("/search", async (c) => {
  const name = c.req.query("name");
  if (!name) {
    return c.json({ error: "missing_param", message: "Provide ?name=example.com" }, 400);
  }

  const domainLower = name.toLowerCase().trim();
  if (!isValidDomain(domainLower)) {
    return c.json({ error: "invalid_domain", message: "Invalid domain name format" }, 400);
  }

  const price = getTldPrice(domainLower);
  const tld = getTldFromDomain(domainLower);

  if (!tld || price === null) {
    return c.json({
      domain: domainLower,
      available: false,
      price_usdc: null,
      tld: tld ? `.${tld}` : null,
      note: "TLD not supported",
    });
  }

  // Check our DB first
  const existing = getDomainByName(domainLower);
  if (existing) {
    return c.json({
      domain: domainLower,
      available: false,
      price_usdc: price,
      tld: `.${tld}`,
      note: "Availability is estimated",
    });
  }

  // Try Njalla get-domain — error means likely available
  let available = true;
  try {
    await njallaGetDomain(domainLower);
    // If no error, domain exists in Njalla → unavailable
    available = false;
  } catch {
    // Error from Njalla = domain not found = available
    available = true;
  }

  return c.json({
    domain: domainLower,
    available,
    price_usdc: available ? price : null,
    tld: `.${tld}`,
    note: "Availability is estimated",
  });
});

// ─── POST /wallet/deposit ───

app.post("/wallet/deposit", requireAuth, (c) => {
  return c.json({
    deposit_address: "0x632881b5f5384e872d8b701dd23f08e63a52faee",
    chain: "Base",
    token: "USDC",
    message: "Send USDC to this address. Contact support to credit your account.",
  });
});

// ─── GET /wallet/balance ───

app.get("/wallet/balance", requireAuth, (c) => {
  const agent = c.get("agent");
  return c.json({
    balance_usdc: Math.round(agent.balance_usdc * 100) / 100,
    agent_id: agent.id,
  });
});

// ─── POST /domains/purchase ───

app.post("/domains/purchase", requireAuth, async (c) => {
  const agent = c.get("agent");
  const agentId = agent.id;

  let body: any;
  try {
    body = await c.req.json();
  } catch {
    return c.json({ error: "invalid_body", message: "Request body must be JSON" }, 400);
  }

  const domain = (body.domain as string)?.toLowerCase()?.trim();
  const years = Math.max(1, Math.min(10, parseInt(body.years ?? "1", 10) || 1));

  if (!domain || !isValidDomain(domain)) {
    return c.json({ error: "invalid_domain", message: "Provide a valid domain name (e.g. example.com)" }, 400);
  }

  const tldPrice = getTldPrice(domain);
  if (tldPrice === null) {
    return c.json({ error: "unsupported_tld", message: "TLD not supported. GET /tlds for the full list." }, 400);
  }

  const totalCost = Math.round(tldPrice * years * 100) / 100;

  // Check if already registered
  if (getDomainByName(domain)) {
    return c.json({ error: "domain_taken", message: `${domain} is already registered` }, 409);
  }

  // Check balance
  const freshAgent = getAgentById(agentId)!;
  if (freshAgent.balance_usdc < totalCost) {
    return c.json(
      {
        error: "insufficient_balance",
        balance_usdc: Math.round(freshAgent.balance_usdc * 100) / 100,
        required_usdc: totalCost,
        message: "Insufficient USDC balance. POST /wallet/deposit for deposit instructions.",
      },
      402
    );
  }

  // Debit balance atomically
  const debited = debitAgent(agentId, totalCost);
  if (!debited) {
    return c.json({ error: "insufficient_balance", message: "Balance check failed" }, 402);
  }

  // Call Njalla register-domain
  let njallaOk = false;
  let expiry: string | null = null;
  try {
    await njallaRegisterDomain(domain);
    njallaOk = true;
    // Estimate expiry as 1 year from now
    const expiryDate = new Date();
    expiryDate.setFullYear(expiryDate.getFullYear() + years);
    expiry = expiryDate.toISOString().split("T")[0];
  } catch (err: any) {
    // Even if Njalla errors, we've already debited — save domain with njalla_registered=0
    // This allows manual follow-up
    console.error(`Njalla registration error for ${domain}:`, err.message);
  }

  // Save domain to DB
  const domainId = uuidv4();
  insertDomain(domainId, agentId, domain, expiry, njallaOk ? 1 : 0);

  // Referral credit (15% of purchase)
  if (freshAgent.referrer_id) {
    const referralAmount = Math.round(totalCost * 0.15 * 100) / 100;
    const referralId = uuidv4();
    insertReferralEarning(referralId, freshAgent.referrer_id, agentId, domain, referralAmount);
    creditAgent(freshAgent.referrer_id, referralAmount);
  }

  const updatedAgent = getAgentById(agentId)!;

  return c.json({
    domain,
    status: njallaOk ? "active" : "pending",
    cost_usdc: totalCost,
    balance_remaining: Math.round(updatedAgent.balance_usdc * 100) / 100,
    message: njallaOk
      ? `${domain} registered successfully`
      : `${domain} purchase recorded. Registration queued — contact support if not active within 24h.`,
  });
});

// ─── GET /domains ───

app.get("/domains", requireAuth, (c) => {
  const agentId = c.get("agentId");
  const domains = getDomainsByAgent(agentId);
  return c.json({
    domains: domains.map((d) => ({
      domain_name: d.domain_name,
      status: d.status,
      expiry: d.expiry,
      registered_at: new Date(d.registered_at * 1000).toISOString(),
    })),
  });
});

// ─── GET /domains/:domain ───

app.get("/domains/:domain", requireAuth, async (c) => {
  const agentId = c.get("agentId");
  const domainName = c.req.param("domain").toLowerCase();

  const dbDomain = getDomainForAgent(domainName, agentId);
  if (!dbDomain) {
    return c.json({ error: "not_found", message: "Domain not found or doesn't belong to you" }, 404);
  }

  // Fetch live info from Njalla
  let njallaInfo: Record<string, any> = {};
  try {
    njallaInfo = await njallaGetDomain(domainName);
  } catch {
    // Non-fatal — return DB record only
  }

  return c.json({
    domain_name: dbDomain.domain_name,
    status: njallaInfo.status ?? dbDomain.status,
    expiry: njallaInfo.expiry ?? dbDomain.expiry,
    registered_at: new Date(dbDomain.registered_at * 1000).toISOString(),
    njalla_registered: dbDomain.njalla_registered === 1,
    locked: njallaInfo.locked ?? null,
    autorenew: njallaInfo.autorenew ?? null,
  });
});

// ─── GET /domains/:domain/records ───

app.get("/domains/:domain/records", requireAuth, async (c) => {
  const agentId = c.get("agentId");
  const domainName = c.req.param("domain").toLowerCase();

  const dbDomain = getDomainForAgent(domainName, agentId);
  if (!dbDomain) {
    return c.json({ error: "not_found", message: "Domain not found or doesn't belong to you" }, 404);
  }

  try {
    const records = await njallaListRecords(domainName);
    return c.json({
      records: records.map((r) => ({
        id: String(r.id),
        name: r.name,
        type: r.type,
        content: r.content,
        ttl: r.ttl,
      })),
    });
  } catch (err: any) {
    return c.json({ error: "fetch_failed", message: err.message }, 500);
  }
});

// ─── POST /domains/:domain/records ───

const VALID_RECORD_TYPES = ["A", "AAAA", "CNAME", "MX", "TXT", "NS"];

app.post("/domains/:domain/records", requireAuth, async (c) => {
  const agentId = c.get("agentId");
  const domainName = c.req.param("domain").toLowerCase();

  const dbDomain = getDomainForAgent(domainName, agentId);
  if (!dbDomain) {
    return c.json({ error: "not_found", message: "Domain not found or doesn't belong to you" }, 404);
  }

  let body: any;
  try {
    body = await c.req.json();
  } catch {
    return c.json({ error: "invalid_body" }, 400);
  }

  const type = (body.type as string)?.toUpperCase();
  const name = (body.name as string) ?? "@";
  const content = body.content as string;
  const ttl = body.ttl ? Number(body.ttl) : 300;

  if (!type || !VALID_RECORD_TYPES.includes(type)) {
    return c.json(
      { error: "invalid_type", message: `type must be one of: ${VALID_RECORD_TYPES.join(", ")}` },
      400
    );
  }
  if (!content) {
    return c.json({ error: "missing_content", message: "content is required" }, 400);
  }

  try {
    const record = await njallaAddRecord(domainName, type, name, content, ttl);
    return c.json({
      id: String(record.id),
      name: record.name ?? name,
      type: record.type ?? type,
      content: record.content ?? content,
      ttl: record.ttl ?? ttl,
    });
  } catch (err: any) {
    return c.json({ error: "add_record_failed", message: err.message }, 500);
  }
});

// ─── DELETE /domains/:domain/records/:id ───

app.delete("/domains/:domain/records/:id", requireAuth, async (c) => {
  const agentId = c.get("agentId");
  const domainName = c.req.param("domain").toLowerCase();
  const recordId = c.req.param("id");

  const dbDomain = getDomainForAgent(domainName, agentId);
  if (!dbDomain) {
    return c.json({ error: "not_found", message: "Domain not found or doesn't belong to you" }, 404);
  }

  try {
    await njallaRemoveRecord(domainName, recordId);
    return c.json({ success: true, message: "Record removed" });
  } catch (err: any) {
    return c.json({ error: "remove_failed", message: err.message }, 500);
  }
});

// ─── PUT /domains/:domain/records/:id ───

app.put("/domains/:domain/records/:id", requireAuth, async (c) => {
  const agentId = c.get("agentId");
  const domainName = c.req.param("domain").toLowerCase();
  const recordId = c.req.param("id");

  const dbDomain = getDomainForAgent(domainName, agentId);
  if (!dbDomain) {
    return c.json({ error: "not_found", message: "Domain not found or doesn't belong to you" }, 404);
  }

  let body: any;
  try {
    body = await c.req.json();
  } catch {
    return c.json({ error: "invalid_body" }, 400);
  }

  const type = (body.type as string)?.toUpperCase();
  const name = body.name as string;
  const content = body.content as string;
  const ttl = body.ttl ? Number(body.ttl) : 300;

  if (!type || !VALID_RECORD_TYPES.includes(type)) {
    return c.json({ error: "invalid_type", message: `type must be one of: ${VALID_RECORD_TYPES.join(", ")}` }, 400);
  }
  if (!name || !content) {
    return c.json({ error: "missing_fields", message: "type, name, and content are required" }, 400);
  }

  // Njalla has no update-record: remove then add
  try {
    await njallaRemoveRecord(domainName, recordId);
  } catch (err: any) {
    return c.json({ error: "remove_failed", message: `Failed to remove old record: ${err.message}` }, 500);
  }

  try {
    const record = await njallaAddRecord(domainName, type, name, content, ttl);
    return c.json({
      id: String(record.id),
      name: record.name ?? name,
      type: record.type ?? type,
      content: record.content ?? content,
      ttl: record.ttl ?? ttl,
    });
  } catch (err: any) {
    return c.json({ error: "add_record_failed", message: `Old record removed but failed to add new: ${err.message}` }, 500);
  }
});

// ─── GET /referrals ───

app.get("/referrals", requireAuth, (c) => {
  const agent = c.get("agent");
  const agentId = agent.id;

  const earnings = getReferralEarnings(agentId);
  const totalEarnings = earnings.reduce((sum, e) => sum + e.amount_usdc, 0);
  const referredCount = getReferralCount(agentId);
  const domainsPurchased = getReferralDomainCount(agentId);

  return c.json({
    referral_code: agentId,
    earnings_usdc: Math.round(totalEarnings * 100) / 100,
    referred_count: referredCount,
    domains_purchased: domainsPurchased,
  });
});

// ─── Start server ───

const port = parseInt(process.env.PORT ?? "3004", 10);
serve({ fetch: app.fetch, port }, (info) => {
  console.log(`Agent Domains API running on http://localhost:${info.port}`);
});

export default app;